+++ install.ps1
@@ -3,8 +3,7 @@ $ErrorActionPreference = 'Stop'
 Remove-Item $PSScriptRoot\mvext-*.vsix -ErrorAction Continue
 vsce.cmd package
 ($ext = (Get-Item $PSScriptRoot\mvext-*.vsix).FullName)
-try {
-	code-insiders.cmd --install-extension $ext
+if (-not $ext) {
+    return 1
 }
-catch {}
 code.cmd --install-extension $ext