+++ src/switchFile.ts
@@ -1,9 +1,10 @@
-import fs from 'fs'
+import fs from 'fs/promises'
 import vscode from 'vscode'
-const { registerCommand, executeCommand } = vscode.commands
 
 export function registerSwitchFile(ctx: vscode.ExtensionContext) {
-  ctx.subscriptions.push(registerCommand('mvext.switchFile', switchFile))
+  ctx.subscriptions.push(
+    vscode.commands.registerCommand('mvext.switchFile', switchFile),
+  )
 }
 
 export async function switchFile() {
@@ -13,52 +14,64 @@ export async function switchFile() {
     return
   }
 
-  let uri: vscode.Uri
-
   if (document.uri.scheme === 'file') {
     let fsPath: string
 
-    switch (document.languageId) {
-      case 'typescript':
-        fsPath = document.fileName.replace(/ts$/, 'js')
-        console.log(fsPath)
-        if (!fs.existsSync(fsPath)) {
-          fsPath = fsPath.replace(/\bsrc\b/, 'out')
-          console.log(fsPath)
-          if (!fs.existsSync(fsPath)) {
-            fsPath = fsPath.replace(/\bout\b/, 'dist')
-            console.log(fsPath)
-            if (!fs.existsSync(fsPath)) {
-              return
-            }
-          }
+    try {
+      switch (document.languageId) {
+        case 'typescript': {
+          // js file
+          fsPath = document.fileName.replace(/ts$/, 'js')
+
+          await fs
+            .access(fsPath)
+            .catch(() => {
+              // out/ dir
+              fsPath = fsPath.replace(/\bsrc\b/, 'out')
+              return fs.access(fsPath)
+            })
+            .catch(() => {
+              // dist/ dir
+              fsPath = fsPath.replace(/\bout\b/, 'dist')
+              return fs.access(fsPath)
+            })
+          break
         }
-        break
-      case 'javascript':
-      case 'css':
-        fsPath = document.fileName.replace(/\.\w+$/, '.html')
-        if (!fs.existsSync(fsPath)) {
-          fsPath = fsPath.slice(0, -1)
-          if (!fs.existsSync(fsPath)) {
-            return
-          }
-        }
-        break
-      case 'html':
-        fsPath = document.fileName.replace(/\.\w+$/, '.css')
-        if (!fs.existsSync(fsPath)) {
+        case 'javascript':
+        case 'css':
+          fsPath = document.fileName.replace(/\.\w+$/, '.html')
+
+          await fs.access(fsPath).catch(() => {
+            // ext .htm
+            fsPath = fsPath.slice(0, -1)
+            return fs.access(fsPath)
+          })
+          break
+        case 'html':
+          fsPath = document.fileName.replace(/\.\w+$/, '.css')
+
+          await fs.access(fsPath).catch(() => {
+            fsPath = fsPath.replace(/css$/, 'js')
+            return fs.access(fsPath)
+          })
+          break
+        default:
           return
-        }
-        break
-      default:
-        return
+      }
+      await vscode.commands.executeCommand(
+        'vscode.open',
+        vscode.Uri.file(fsPath),
+      )
+    } catch {
+      return
     }
-    uri = vscode.Uri.file(fsPath)
   } else {
+    let uri: vscode.Uri
+
     switch (document.languageId) {
       case 'typescript':
         uri = document.uri.with({
-          path: document.uri.path.replace(/ts$/, 'js'),
+          path: '/__tests__' + document.uri.path.replace(/ts$/, 'test.ts'),
         })
         break
       case 'javascript':
@@ -75,7 +88,6 @@ export async function switchFile() {
       default:
         return
     }
+    await vscode.commands.executeCommand('vscode.open', uri)
   }
-  console.log(uri.fsPath)
-  await executeCommand('vscode.open', uri)
 }