+++ src/test/suite/extension.test.ts
@@ -1,20 +1,14 @@
-// import vscode from 'vscode'
+import vscode from 'vscode'
 import strict from 'assert/strict'
 import { describe, it } from 'mocha'
 
 describe('vscode', function () {
-  it('work', function () {
-    // try {
-    //   const inputVal = await vscode.window.showInputBox({
-    //     title: 'test',
-    //     prompt: 'I will show your inputs',
-    //     value: 'guess?',
-    //   })
-    //   console.log(inputVal)
-    // } catch (err) {
-    //   console.log('vsc err')
-    //   console.error(err)
-    // }
+  it('work', async function () {
+    await vscode.window
+      .showInformationMessage('Hello World')
+      .then(undefined, (err) => {
+        strict.ok(err instanceof Error)
+      })
     strict.equal(-1, [1, 2, 3].indexOf(5))
     strict.equal(-1, [1, 2, 3].indexOf(0))
   })