+++ src/utils/getExtConfig.ts
@@ -1,0 +1,5 @@
+import vscode from 'vscode'
+
+export function cfgEvalWithSelection() {
+  return vscode.workspace.getConfiguration('mvext.evalWithSelection')
+}